requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.2
django-bootstrap5==24.3
django-cors-headers==4.5.0
djangorestframework==3.15.2
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
